+++ tetris/build.gradle
@@ -1,11 +1,11 @@
 plugins {
     id 'java'
     id 'application'
-    id 'org.openjfx.javafxplugin' version '0.0.12'
+    id 'org.openjfx.javafxplugin' version '0.0.13'
 }
 
 javafx {
-    version = "20"
+    version = '17'
     modules = [ 'javafx.controls', 'javafx.fxml' ]
 }
