--- tetris/src/main/java/ru/nsu/fit/vinter/tetris/core/model/Point.java
@@ -1,15 +1,0 @@
-package ru.nsu.fit.vinter.tetris.core.model;
-
-public class Point {
-    private int x;
-    private int y;
-
-    public int getX() {
-        return x;
-    }
-    public void subX() { this.x--; }
-    public int getY() {
-        return y;
-    }
-    public void subY() { this.y--; }
-}
