+++ tetris/src/main/java/ru/nsu/fit/vinter/tetris/core/model/Tetris.java
@@ -1,0 +1,4 @@
+package ru.nsu.fit.vinter.tetris.core.model;
+
+public class Tetris {
+}
