+++ tetris/src/main/java/ru/nsu/fit/vinter/tetris/executable/Main.java
@@ -1,11 +1,11 @@
 package ru.nsu.fit.vinter.tetris.executable;
 
 import javafx.application.Application;
-import ru.nsu.fit.vinter.tetris.core.presenter.Presenter;
+import ru.nsu.fit.vinter.tetris.core.controller.Controller;
 
 public class Main {
     public static void main(String[] args) {
-        Presenter presenter = new Presenter();
-        Application.launch(Presenter.class);
+        Controller controller = new Controller();
+        Application.launch(Controller.class);
     }
 }
