+++ tetris/src/main/resources/highScores.fxml
@@ -1,20 +1,28 @@
 <?xml version="1.0" encoding="UTF-8"?>
 
+<?import java.lang.*?>
+<?import java.util.*?>
+<?import javafx.scene.*?>
 <?import javafx.scene.control.*?>
 <?import javafx.scene.layout.*?>
-<?import javafx.scene.text.*?>
 
-<AnchorPane prefHeight="800.0" prefWidth="600.0" xmlns="http://javafx.com/javafx/17.0.2-ea" xmlns:fx="http://javafx.com/fxml/1" fx:controller="ru.nsu.fit.vinter.tetris.core.presenter.Presenter">
-      <Button fx:id="exitButton2" alignment="CENTER" contentDisplay="CENTER" layoutX="230.0" layoutY="600.0" mnemonicParsing="false" onMouseClicked="#getBackButtonCLicked" prefHeight="40.0" prefWidth="140.0" text="Get Back" />
-      <GridPane layoutX="200.0" layoutY="200.0" minHeight="50.0" minWidth="100.0">
+
+<Pane prefHeight="800.0" prefWidth="600.0"
+      xmlns="http://javafx.com/javafx/17.0.2-ea"
+      xmlns:fx="http://javafx.com/fxml/1"
+      fx:controller="ru.nsu.fit.vinter.tetris.core.controller.Controller">
+    <Button fx:id="exitButton2" alignment="CENTER" contentDisplay="CENTER"
+            layoutX="230.0" layoutY="600.0" mnemonicParsing="false"
+            onMouseClicked="#getBackButtonCLicked" prefHeight="40.0" prefWidth="140.0" text="Get Back" />
+    <GridPane layoutX="200.0" layoutY="200.0" minHeight="50.0" minWidth="100.0">
         <columnConstraints>
-          <ColumnConstraints hgrow="SOMETIMES" maxWidth="100.0" minWidth="10.0" prefWidth="200.0" />
-          <ColumnConstraints hgrow="SOMETIMES" maxWidth="100.0" minWidth="10.0" prefWidth="100.0" />
+            <ColumnConstraints hgrow="SOMETIMES" maxWidth="100.0" minWidth="10.0" prefWidth="200.0" />
+            <ColumnConstraints hgrow="SOMETIMES" maxWidth="100.0" minWidth="10.0" prefWidth="100.0" />
         </columnConstraints>
         <rowConstraints>
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
-          <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
+            <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
             <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
             <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
             <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
@@ -23,10 +31,5 @@
             <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
             <RowConstraints minHeight="10.0" prefHeight="30.0" vgrow="SOMETIMES" />
         </rowConstraints>
-      </GridPane>
-      <TextArea layoutX="225.0" layoutY="100.0" prefHeight="70.0" prefWidth="150.0" text="High Scores:">
-         <font>
-            <Font size="20.0" />
-         </font>
-      </TextArea>
-</AnchorPane>
+    </GridPane>
+</Pane>
