--- tetris/src/main/resources/highScores.txt
@@ -1,10 +1,0 @@
-1 ---
-2 ---
-3 ---
-4 ---
-5 ---
-6 ---
-7 ---
-8 ---
-9 ---
-10 ---
